+++ Cargo.toml
@@ -32,7 +32,7 @@ required-features = [ "bench", "proptest" ]
 async-trait = "0.1.53"
 cli-batteries = { version = "0.1.5", features = [ "prometheus", "metered_allocator" ] }
 # ethers = { git = "https://github.com/gakonst/ethers-rs.git", branch = "master", rev = "4d99165", features = [ "ws", "ipc", ]  }
-ethers = { git = "https://github.com/recmo/ethers-rs.git", rev = "93e8e52", features = [ "ws", "ipc", ]  }
+ethers = { git = "https://github.com/recmo/ethers-rs.git", rev = "93e8e52", features = [ "ws", "ipc", "openssl" ]  }
 eyre = "0.6"
 futures = "0.3"
 hyper = { version = "^0.14.17", features = [ "server", "tcp", "http1", "http2" ] }
@@ -47,7 +47,7 @@ tokio = { version = "1.17", features = [ "signal", "macros", "rt", "sync", "time
 tracing = "0.1"
 tracing-futures = "0.2"
 url = "2.2"
-sqlx = { version = "0.5", features = [ "runtime-tokio-native-tls", "any", "postgres" ] }
+sqlx = { version = "0.5", features = [ "runtime-tokio-native-tls", "any" ] }
 chrono = "0.4.19"
 reqwest = "0.11.10"
 ruint = { version = "1.3", features = ["primitive-types", "sqlx"] }
