+++ build.rs
@@ -1,3 +1,5 @@
 fn main() {
     cli_batteries::build_rs().unwrap();
+
+    println!("cargo:rerun-if-changed=schemas");
 }
