+++ schemas/database/Readme.md
@@ -1,0 +1,7 @@
+# Database migration files
+
+These migration scripts are statically linked into the application. They must
+be names `<VERSION>_<DESCRIPTION>.sql` for simple migrations. Complex migrations
+can be done using `.up.sql` and `.down.sql` scripts.
+
+Migrations are tracked and executed using `sqlx`.
