+++ src/database/mod.rs
@@ -1,0 +1,162 @@
+use crate::app::Hash;
+use eyre::{eyre, Context, Result};
+use ruint::{aliases::U256, uint};
+use sqlx::{
+    any::AnyKind,
+    migrate::{Migrate, Migrator},
+    pool::PoolOptions,
+    Any, Executor, Pool, Row,
+};
+use structopt::StructOpt;
+use tracing::{debug, error, info, instrument, warn};
+use url::Url;
+
+// Statically link in migration files
+static MIGRATOR: Migrator = sqlx::migrate!("schemas/database");
+
+#[derive(Clone, Debug, PartialEq, Eq, StructOpt)]
+pub struct Options {
+    /// Database server connection string.
+    /// Example: `postgres://user:password@localhost:5432/database`
+    /// Sqlite file: ``
+    /// In memory DB: `sqlite::memory:`
+    #[structopt(
+        long,
+        env,
+        default_value = "postgres://postgres:password@localhost/test"
+    )]
+    pub database: Url,
+
+    /// Allow creation or migration of the database schema.
+    #[structopt(long)]
+    pub database_migrate: bool,
+
+    /// Maximum number of connections in the database connection pool
+    #[structopt(long, env, default_value = "10")]
+    pub database_max_connections: u32,
+}
+
+pub struct Database {
+    pool: Pool<Any>,
+}
+
+impl Database {
+    #[instrument(skip_all)]
+    pub async fn new(options: Options) -> Result<Self> {
+        info!(url = %&options.database, "Connecting to database");
+
+        // // Create database if requested and does not exist
+        // if options.database_migrate &&
+        // !Any::database_exists(options.database.as_str()).await? {
+        //     warn!(url = %&options.database, "Database does not exist, creating
+        // database");     Any::create_database(options.database.as_str()).
+        // await?; }
+
+        // Create a connection pool
+        let pool = PoolOptions::<Any>::new()
+            .max_connections(options.database_max_connections)
+            .connect(options.database.as_str())
+            .await
+            .wrap_err("error connecting to database")?;
+
+        // Log DB version to test connection.
+        let sql = match pool.any_kind() {
+            #[cfg(feature = "sqlite")]
+            AnyKind::Sqlite => "sqlite_version() || ' ' || sqlite_source_id()",
+
+            #[cfg(feature = "postgres")]
+            AnyKind::Postgres => "version()",
+
+            // Depending on compilation flags there may be more patterns.
+            #[allow(unreachable_patterns)]
+            _ => "'unknown'",
+        };
+        let version = pool
+            .fetch_one(format!("SELECT {sql};", sql = sql).as_str())
+            .await
+            .wrap_err("error getting database version")?
+            .get::<String, _>(0);
+        info!(url = %&options.database, kind = ?pool.any_kind(), ?version, "Connected to database");
+
+        // Run migrations if requested.
+        let latest = MIGRATOR.migrations.last().unwrap().version;
+        if options.database_migrate {
+            info!(url = %&options.database, "Running migrations");
+            MIGRATOR.run(&pool).await?;
+        }
+        // Validate database schema version
+        #[allow(deprecated)] // HACK: No good alternative to `version()`?
+        if let Some((version, dirty)) = pool.acquire().await?.version().await? {
+            if dirty {
+                error!(
+                    url = %&options.database,
+                    version,
+                    expected = latest,
+                    "Database is in incomplete migration state.",
+                );
+                return Err(eyre!("Database is in incomplete migration state."));
+            } else if version < latest {
+                error!(
+                    url = %&options.database,
+                    version,
+                    expected = latest,
+                    "Database is not up to date, try rerunning with --database-migrate",
+                );
+                return Err(eyre!(
+                    "Database is not up to date, try rerunning with --database-migrate"
+                ));
+            } else if version > latest {
+                error!(
+                    url = %&options.database,
+                    version,
+                    latest,
+                    "Database version is newer than this version of the software, please update.",
+                );
+                return Err(eyre!(
+                    "Database version is newer than this version of the software, please update."
+                ));
+            }
+            info!(
+                url = %&options.database,
+                version,
+                latest,
+                "Database version is up to date.",
+            );
+        } else {
+            error!(url = %&options.database, "Could not get database version");
+            return Err(eyre!(
+                "Could not get database version."
+            ));
+        }
+
+        Ok(Self { pool })
+    }
+
+    pub async fn read(&self, index: usize) -> Result<Hash> {
+        self.pool
+            .execute(sqlx::query(
+                r#"CREATE TABLE IF NOT EXISTS hashes (
+                id SERIAL PRIMARY KEY,
+                hash TEXT NOT NULL
+            );"#,
+            ))
+            .await?;
+
+        let value = uint!(0x12356_U256);
+
+        self.pool
+            .execute(sqlx::query(r#"INSERT INTO hashes ( hash ) VALUES ( $1 );"#).bind(value))
+            .await?;
+
+        let rows = self
+            .pool
+            .fetch_all(sqlx::query(r#"SELECT hash FROM hashes;"#))
+            .await?;
+        for row in rows {
+            let hash = row.get::<U256, _>(0);
+            info!(hash = ?hash, "Read hash");
+        }
+
+        Ok(Hash::default())
+    }
+}
